Further generalize EpochBatchIterator and move iterators into new file
